Restore old 1.16 behavior when deploying a service
